updated static files, add uml diagram modified flask app
