--- Designs/templates/index.html
@@ -1,29 +1,0 @@
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Dash Board</title>
-    <link rel="stylesheet" href="../css/style.css">
-</head>
-
-<body class="dash-container">
-    <div class="header">
-        <a class="header-a">Yummy Recipes v1.0 </a>
-        <span class="recover_pass"> <a href="#" class="signout">SignOut</a></span>
-    </div>
-    <nav>
-        <div>
-            <ul>
-                <li><a class="dash-head">Dash Board </a></li>
-                <li><a href="/category">Add Recipes Category</a></li>
-                <li><a href="/recipes">View Recipes </a></li>
-            </ul>
-        </div>
-
-    </nav>
-
-</body>
-
-</html>
